primitives - split `logging` and `serde` modules in `util`
